feat: Enhance Dockerfile for static gptfdisk build
- Added pkgconfig support to streamline dependencies
- Updated build flags to create a static binary, reducing size by approximately 20MB
- Improved installation process for gptfdisk

This results in the gdisk binary growing from 500Kb to 9Mb, but makes us avoid having to ship the full libstdc++ lib which is about 20Mb. Previous results to try to strip the lib produced panics so its the only way to reduce the global size.

For now only sgdisk needs that lib.

Signed-off-by: Itxaka <itxaka@kairos.io>