Fix systemd-home modules in pam
Seems liek systemd-home was calling homed on ssh command run, which
tried to get the home dir from systemd-homed. But we dont build homed so
it was returning a noisy error which would confuse the exit status.
Later down the line, pam does the proper thing and cds into the home
user directory, so this was adding only noise and confusion.

Drop the systemd-home entries in pam to avoid this.

This also install bashrc to the proper bash.bashrc location

Signed-off-by: Itxaka <itxaka@kairos.io>