fix: Update multipath-tools and related dependencies
- Bump MULTIPATH_TOOLS_VERSION from 0.11.3 to 0.13.0
- Bump LVM2_VERSION from 2.03.35 to 2.03.37
- Bump URCU_VERSION from 0.15.2 to 0.15.3
- Adjust build commands for multipath-tools to ensure proper installation

Patch multipathd to avoid a coredump on cleanup.

Seems like while glibc is more careful about freeing a thread when the pthread_join is called, musl is not and calls to an already dereferenced pointer which causes a coredump.

checking the code seems like urcu already calls the pthread_join when call_rcu_data_free is called, so we are doing a double free on the same thread. Fix this by not calling the pthread_join again.

This patch was also sent to upstream to multipath directly as it fixes a real issue which also shows on other musl distros like alpine.

These updates address compatibility issues and improve functionality during cleanup processes.

Signed-off-by: Itxaka <itxaka@kairos.io>